requirements.txt
PyTextNow ~= 1.1.9
discord.py ~= 1.4.3
pymongo~= 4.1.1
python-dotenv~=0.20.0
dnspython ~=2.2.1